requirements.txt
numpy
sounddevice
flask
python-rtmidi  # optional — needed only for synth_realtime.py --midi mode
